+++ srcs/validate_args.hpp
@@ -3,7 +3,7 @@
 
 #include <string>
 
-int num_of_digit(int num);
+unsigned long num_of_digit(int num);
 int validate_port(char *port);
 void validate_password(std::string password);
 int validate_args(int argc, char *argv[]);
